12.23 teszt 28s es köridő és SC követés sebességszabáylozással működik.

+++ RobonAUT-control/Core/Inc/main.h
@@ -143,7 +143,7 @@ extern uint8_t swState[];
 extern uint8_t fromPC[];
 
 extern float v;
-extern int32_t v_ref;
+extern float v_ref;
 //um/s
 
 /* USER CODE END Private defines */
